+++ src/filters/parser/DiracSplitter/DiracSplitter.cpp
@@ -137,7 +137,7 @@ HRESULT CDiracSplitterFilter::CreateOutputs(IAsyncReader* pAsyncReader)
 
 bool CDiracSplitterFilter::DemuxInit()
 {
-	SetThreadName(-1, "CDiracSplitterFilter");
+	SetThreadName((DWORD)-1, "CDiracSplitterFilter");
     if(!m_pFile) return(false);
 
     // TODO
