+++ src/filters/renderer/VideoRenderers/VMR7AllocatorPresenter.cpp
@@ -195,7 +195,7 @@ STDMETHODIMP CVMR7AllocatorPresenter::FreeSurface(DWORD_PTR dwUserID)
 
 STDMETHODIMP CVMR7AllocatorPresenter::PrepareSurface(DWORD_PTR dwUserID, IDirectDrawSurface7* lpSurface, DWORD dwSurfaceFlags)
 {
-	SetThreadName(-1, "CVMR7AllocatorPresenter");
+	SetThreadName((DWORD)-1, "CVMR7AllocatorPresenter");
 
     if(!lpSurface)
         return E_POINTER;
